+++ api/src/main/java/org/mandulis/mts/group/GroupRequest.java
@@ -3,15 +3,13 @@ package org.mandulis.mts.group;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import lombok.*;
 
 
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
+@Getter
 public class GroupRequest {
     @NotBlank(message = "Group name cannot be blank")
     @NotNull
