+++ api/src/main/java/org/mandulis/mts/storage/MinioStorageService.java
@@ -1,4 +1,4 @@
-package org.mandulis.mts.service.storage;
+package org.mandulis.mts.storage;
 
 import io.minio.BucketExistsArgs;
 import io.minio.MinioClient;
