+++ api/src/test/java/org/mandulis/mts/controller/AttachmentControllerIT.java
@@ -4,8 +4,8 @@ import io.minio.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.mandulis.mts.entity.Attachment;
-import org.mandulis.mts.repository.AttachmentRepository;
+import org.mandulis.mts.attachment.Attachment;
+import org.mandulis.mts.attachment.AttachmentRepository;
 import org.mandulis.mts.utils.DatabaseHelper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
