+++ api/src/test/java/org/mandulis/mts/service/AttachmentValidationServiceTest.java
@@ -1,6 +1,7 @@
 package org.mandulis.mts.service;
 
 import org.junit.jupiter.api.Test;
+import org.mandulis.mts.attachment.AttachmentValidationService;
 import org.mandulis.mts.dto.AttachmentDto;
 import org.mandulis.mts.exception.AttachmentWithoutNameException;
 import org.springframework.mock.web.MockMultipartFile;
